+++ src/main/java/roomescape/auth/token/jwt/JwtProperties.java
@@ -1,4 +1,4 @@
-package roomescape.auth.session.jwt;
+package roomescape.auth.token.jwt;
 
 import org.springframework.stereotype.Component;
