+++ src/main/java/roomescape/global/configuration/WebConfiguration.java
@@ -5,8 +5,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.auth.session.cookie.CookieResolver;
-import roomescape.auth.session.jwt.JwtResolver;
+import roomescape.auth.token.cookie.CookieResolver;
+import roomescape.auth.token.jwt.JwtResolver;
 import roomescape.global.interceptor.AuthInterceptor;
 import roomescape.global.resolver.MemberArgumentResolver;
 import roomescape.member.MemberService;
