+++ src/main/java/roomescape/global/interceptor/AuthInterceptor.java
@@ -4,9 +4,9 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
-import roomescape.auth.session.cookie.CookieResolver;
-import roomescape.auth.session.jwt.JwtProperties;
-import roomescape.auth.session.jwt.JwtResolver;
+import roomescape.auth.token.cookie.CookieResolver;
+import roomescape.auth.token.jwt.JwtProperties;
+import roomescape.auth.token.jwt.JwtResolver;
 
 @Component
 public class AuthInterceptor implements HandlerInterceptor {
