+++ src/test/java/roomescape/auth/AuthServiceTest.java
@@ -10,7 +10,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.auth.session.jwt.JwtProperties;
+import roomescape.auth.token.jwt.JwtProperties;
 import roomescape.member.Member;
 
 @SpringBootTest(webEnvironment = WebEnvironment.NONE)
