+++ src/test/java/roomescape/auth/token/cookie/CookieProviderTest.java
@@ -1,14 +1,14 @@
-package roomescape.auth.session.cookie;
+package roomescape.auth.token.cookie;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static roomescape.auth.session.cookie.CookieProvider.PATH;
+import static roomescape.auth.token.cookie.CookieProvider.PATH;
 
 import java.time.Duration;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.ResponseCookie;
-import roomescape.auth.session.jwt.JwtProperties;
+import roomescape.auth.token.jwt.JwtProperties;
 
 class CookieProviderTest {
